Clone to avoid sharing mutations from `transform` between siblings in `traverse`
Summary:
The `Transform` trait implementation for a particular type is split between two functions:
- transform
- traverse

The `transform` function does the following:
- clones the received pass (for use with the bottom up transform function, in the last step)
- applies the `on_[...]_top_down` transform function which may mutate the pass
- traverses its children with the possibly mutabled pass
- applies the `on_[...]_bottom_up` transform function to the cloned, original pass

The `traverse` function is there to apply a `transform` to each element of a type.

For a tuple, the implementation simply calls `transform` on each element:

```
fn traverse(&mut self, env: &Env, pass: &mut (impl Pass + Clone)) {
        self.0.transform(env, &mut pass.clone());
        self.1.transform(env, &mut pass.clone())
}
```

Here, we have to clone the pass on each call to `transform` since, as noted above, we may mutate the pass in that call and we want to ensure any mutations are not shared between siblings.

At the moment, we are not ensuring this for types with derived implementations e.g. this is the generated code for `UsingStmt`:

```
fn traverse(&mut self, env: &Env, pass: &mut (impl Pass + Clone)) {
        match self {
            UsingStmt {
                is_block_scoped: ref mut __binding_0,
                has_await: ref mut __binding_1,
                exprs: ref mut __binding_2,
                block: ref mut __binding_3,
            } => {
                {
                    __binding_0.transform(env, pass)
                }
                {
                    __binding_1.transform(env, pass)
                }
                {
                    __binding_2.transform(env, pass)
                }
                { __binding_3.transform(env, pass) }
            }
        }
}
```

This diff modifies codegen to ensure we clone the pass and avoid sharing mutation across siblings.

Reviewed By: shayne-fletcher

Differential Revision: D50694675

fbshipit-source-id: bc50265403ccc98b8c785e6662afbfd4766b10e4