Do not pessimize inlining in Switch/SSwitch
Summary:
Switch and SSwitch bytecodes have multiple destinations. The current IR model
cannot represent that with a single instruction, so this was implemented with
binding adresses to all destinations, loading them from a table and jumping to
them. This required all inlined frames to be spilled.

Change that into a two step system. Load the bound address. If we are in an
inlined frame, use the recently introduced sideExitFromInlined logic that will
materialize only the current frame and set up return to the current inline
context, pessimizing only the side exiting frame instead of all frames. If we
are not in an inlined frame, just JmpExit to that address (renamed from
JmpSSwitchDest).

Previously, JmpSwitchDest with a known input would get simplified to
ReqBindJmp. We now simplify LdSwitchDest to LdBindAddr and add a similar
optimization for LdSSwitchDest. We also simplify a combination of LdBindAddr
and JmpExit to ReqBindJmp, preserving the previous optimizations in non-inlined
context.

A combination of LdSwitchDest and JmpExit now creates load+jmpr vs previous
JmpSwitchDest creating jmpm. Fuse these instructions together on the vasm level.

This should result in very minor performance gains and code size reduction. It
is a step towards killing spillInlinedFrams().

Reviewed By: ricklavoie

Differential Revision: D50960162

fbshipit-source-id: a3271962e9a0cc87784ad6d7a30496394d1d9416