handle correctly closures generated by methods in module level traits
Summary:
Closures generated by methods in module level traits must belong to the module where the trait was defined, not the module where the trait is used.

This diffs ensures that HHBBC tags with `requiresFromOriginalModule` the `invoke`methods in closure classes that are copied across module boundaries to other units during HHBBC inlining traits.

Reviewed By: oulgen

Differential Revision: D50936657

fbshipit-source-id: 19e5cf605f546029be5519d011f5123c264c1130