Fix vtable slot calculation in HHBBC
Summary:
The changes to use ClassGraph introduced a big with vtable
slot calculations. If a class becomes "conservative", we drop the
subclass list. The problem is, if the class is a child in a later
round, we'll need that list to calculate interface conflict sets. This
causes some conflicts to not be reported.

Calculate the interface conflicts in the flatten classes pass instead.

Reviewed By: oulgen

Differential Revision: D50966842

fbshipit-source-id: ff060f50fe037171d86452a220673a749586c256