Add immediate to AsTypeHintException bytecode
Summary:
To control which exception is raised. Currently `as` raises an TypeHintException, which is something we could reasonably expect a developer to catch. We want to re-use `as` for enforcement of typed locals, but there catching the exception should be limited to more exotic circumstances, and so we would rather throw the same recoverable error that parameter/return enforcement does.

Will plumb this in in a subsequent diff

Reviewed By: rjbailey

Differential Revision: D50598944

fbshipit-source-id: bdb233cb3a3dea9b5b0793c00e9c419da7b33a7a