codegen isPrimitive
Summary:
The `isinstance(t, EnumTypeInfo)` condition in `isPrimitive` always evaluates to `False` because `t` in that case is a `labmda`. But we can't call the lambda during class creation time; can only safely call it when done.

If we want enums to be treated as primitives (eagerly set from internal data tuple), then one solution is replacing `isPrimitive` with a codegen'd bool in class spec.

#buildmore

Reviewed By: nanshu

Differential Revision: D50523077

fbshipit-source-id: 9c05deb9940ad8e091fccf96936dadfcdc1080ea