fix up telemetry
Summary:
Previosuly we were looking at the decl hit rate as # decls fetched / # names, but this is erroneous since not all names are fetched. eg. if you add
a completely new class, then of course we won't have the decl for it. This diff updates the telemetry to use the # decls we know are in the naming table
and thus should be fetchable as the denominator. This is probably going to be a temporary thing as once we remove the old version, it's a bit inefficient
to fetch the decl hashes from the naming table purely for telemetry reasons.

Reviewed By: nt591

Differential Revision: D51031726

fbshipit-source-id: cc8d00071df43b2a665c002cd09f039be955b2ce