Allow to dump some TASTs while typechecking with a new map-reduce service
Summary:
With this diff, passing `--config dump_tasts=file_list.txt` to hh will dump the TASTs of the files listed in `file_list.txt` into `/tmp/hh_{server,distc}/tasts`.

Design choices:
* This is implemented with a new map-reducer `Tast_collector`
Why: It needs to work with both hh_server and hh_distc (to be able to compare). Typechecking might therefore be done on remote workers, so we need to gather TASTs to the master so it can dump them.
* TASTs are gathered as plain strings.
Why: If we stored them as Tast.t, there is no way that I know of to serialize with the exact same string from both Rust and Ocaml.

Differential Revision: D50975307

fbshipit-source-id: 27fe451556b4d12d8784f51e7feb27e1fe1d6026