Use create_catch_block() instead of EndCatchMode::CatchMode::SideExit
Summary:
The EndCatchMode::CatchMode::SideExit value is used to mark catch blocks that
have a special side-exit in case of an InvalidSetMException. But it is not
really needed, because if we actually reach EndCatch, it means we did not hit
the side-exit.

Indeed, the flag is checked only by DCE, but it is never true in blocks that
start with BeginCatch and end with EndCatch. But if we somehow managed to
optimize away UnwindCheckSideExit and turned such catch block into a single
actual block, the DCE optimization would still be valid, so it is safe to
remove that check from there.

The only other place that checks for it is endCatchFromInlined, where it is
also unnecessary.

So, kill the flag and replace the custom catch block construction with a call
to create_catch_block(). This replaces a call to legacy spillInlinedFrames()
(if an actual exception happens) with the new endCatchFromInlined logic,
reducing the size of the catch block, especially when inlining.

Reviewed By: ricklavoie

Differential Revision: D51052910

fbshipit-source-id: 54de7c4eba90ab3c1b21627a6598b17abda5c52b