Adding ext_decl exposing the rust hack parser API to hack 1/N
Summary:
Adding a wrapper class around hackc DeclHolder that is exposed to hack code. This allows some file parsing which should be significantly faster than autoloading + compiling + using reflection.

Note that this diff adds a parsePath method that is not yet implemented. It will be implemented in a following diff along with caching and cache invalidation logic.

Usage:
```
$instance = FileDecls::parseText('<?hh .... ');
if ($instance->getError() == '') {
  var_dump($instance->getFile());
}
```

Reviewed By: nt591

Differential Revision: D50666662

fbshipit-source-id: d51a6b4bd45a4831aee19d56aec063b078e73481