Use type dynamic for result of static method call through string
Summary: If a static method is accessed through a string rather than `dynamic`, we produce an error, and internally generate a fresh type variable that gets solved to anything that fits, typically `nothing`. If hidden by a fixme (error code 4026) this leads to unsoundness (types that don't tell the truth). Better to produce the type `dynamic`.

Reviewed By: madgen

Differential Revision: D51107248

fbshipit-source-id: ee84d557cbc457c960f725b53d77ac8fa9677622