Attempt to give type to array access even when nullable
Summary: Currently we just reject array access through a nullable, unless surrounded by null coalesce operator (e.g. `$a[2] ?? $b`), and produce a error type variable as the type of the result, which typically gets solved to `nothing`. Better to continue typing, and detect downstream errors from the element type of the array; also, this makes a FIXME of the array access just ignore the nullability but avoid further unsound inference.

Reviewed By: madgen

Differential Revision: D51114177

fbshipit-source-id: 2b98629b9669c746f0b3f6860435ea216dabe604