use 'is_rootcanal_signed' from the config
Summary:
'is_rootcanal_signed' is set in the config based on who issues the given cert.
but this field in not used anywhere in the proxygen code, yet.

let's read this field to set 'isProdCASigned'. it'll be used to decide if the
cert can be fetched directly from ProdCA.

"ProdCA" is used because its the preferred name vs "rootcanal".

Reviewed By: zalecodez

Differential Revision: D51176576

fbshipit-source-id: 5827b2902d1a7b9d08067a14d723df78cef5cc0d