Type echo using an hhi signature
Summary:
The type checker has special logic for the pseudo-function `echo`. Instead let's define a signature for it in an hhi file, to eliminate the use of `Tany` in the typing logic, and to simplify the type-checker.
```
function echo(readonly arraykey ...$args)[io]: void;
```
The above is the "ideal" signature. Unfortunately, `readonly` is rejected on varargs by the parser (so typing logic adds it) and `io` is not available as an effect in `hhi` files (so typing logic adds it).

Reviewed By: mheiber, compositor

Differential Revision: D51026539

fbshipit-source-id: 3fefee1277e1ce25f503b4640f0d196f8330da2a