Type sequences of top-level statements
Summary:
Before this change, we typed each top-level statement by itself, discarding the typing environment after each.

With this diff, we propagate a typing environment for top-level statements.

This enabled us to have the expected behavior for code like:

```
<?hh
$x = new Foo();
$x->meth();
```

See the snapshot test for the change in behavior.

The goal of this change is to enable language server features to work well for top-level statements (including type-guided hovers and completions) because top-level statements are used in Hack notebooks.

Even though the goal is related to notebooks, this change affects all users of Hack, but hopefully in ways that either don't matter or are helpful.

This diff has no effect on our parse error for top-level statements:

```
Parsing[1002] Hack does not support top level statements. Use the __EntryPoint attribute on a function instead [1]

/home/mheiber/test.php:9:1
    7 |
    8 | $x = new Foo();
[1] 9 | $y = $x->bar();

2 errors found.
```

Differential Revision: D51028939

fbshipit-source-id: 988f4c3b6365f65ba9ac568047a2451b09c6338a