add --skip-gen option (and minor refactoring of args parsing logic)
Summary:
The `--skip-gen` option allows the compiler to complete all steps prior to code generation (including IDL parsing, mutation and validation), but skip the code generation itself.

This is particularly useful when looking to validate Thrift IDL (eg., run over all thrift files in a given repo), as it saves significant time. In local devserver runs, validating all thrift files in fbcode/, we see a ~23x reduction in time when skipping generation (in that case, of the `ast` generator):  {F1143475503}

*Version 2*
While addressing the review comment (to enforce that either `--gen` or `--skip-gen` is used, but not both), I took the opportunity to slightly refactor the argument parsing logic by moving small parts in separate (internal) methods.

Reviewed By: iahs

Differential Revision: D50462929

fbshipit-source-id: 06902f5dc93afdfb9724955de7d117cac3ff0819