clarify direct-decl-parser hashes
Summary:
There was an incorrect comment about position-free hashes, which I fixed.

I also removed a function `Direct_decl_parser.decls_hash` which feels a bit dangerous -- it will either be the same as pfh_file_hash or different, depending on whether you have the exact same decls in the exact same order, whether you've removed or not removed duplicates in the same way, whether you've removed stdlib or not. (I started by trying to document it, but it felt too subtle, and since it's not being used then I thought it easier to just delete).

Reviewed By: rjbailey

Differential Revision: D51124424

fbshipit-source-id: a31ee12bd3f74aa535c70a9cc7c8612800ef76b8