Revive EndCatch Teardown Optimizations
Summary:
Add a new mode to EndCatch that customizes teardown in the unwinder. Specifically, we explicitly generate decrefs of stack and frame slots when there are only a small number of refcounted elements and tell the unwinder to not perform any decrefs.

In addition, since we now know that all the local slots and stack slots are now dead, the memory effects of EndCatch can indicate this allowing stores in such catch blocks to be eliminated.

Reviewed By: jano

Differential Revision: D48726090

fbshipit-source-id: 69e3753efde715f909325d35979fd430ab1c785b