Closures of an uninstantiable class should not be instantiable in HHBBC (v2)
Summary:
If a class is uninstantiable, then any closures within the
class should also be uninstantiable. Unfortunately we weren't getting
the logic right, leading to conditions where we could assert.

Fix this by detecting it in the FlattenJob, where it belongs.

The first attempt failed because there's some situations where a
closure may end up on a different shard than it's context class, and
thus never realize the context class is uninstantiable.

Reviewed By: mdko

Differential Revision: D51235728

fbshipit-source-id: f96c238b8e07db6a6ae24127a778759dfbe1b32c