Produce sensible type even when field doesn't exist in shape type
Summary: For shape accesses, such as `$s['a']`, if field `'a'` doesn't exist in the shape type of `$s`, instead of producing an error type variable (which typically gets solved to `nothing`), use the "`...`" open type of the shape to give a sound typing to the expression. This ensures that a `HH_FIXME[4108]` does not break True Types.

Reviewed By: madgen

Differential Revision: D51160083

fbshipit-source-id: 77cd06e61b102000a33ea7886717aabbbb4fdadf