Use real path rather than default path when typing top-level statements
Summary:
## What

iuc this change is either a no-op from the user's point of view or removes an edge case we didn't find yet

## How

For typing top-level statements, find the path using a position from the AST. We do the same elsewhere for top-level functions.

## Why

We were using a `Relative_path.Default` when making an environment for top-level statements before since top-level statements were last added in 2018 (D9702794):

https://www.internalfb.com/code/fbsource/[ee1b617b3809f1db2e4e4fd2cbcc86a7ec1d9280]/fbcode/hphp/hack/src/typing/typing.ml?lines=7129-7130

However, parts of the codebase have unknown or strange behavior the presence of a `Relative_path.Default`, so I think it's safer to use a real position.

Differential Revision: D51160021

fbshipit-source-id: 83fe09dc0281e0cb8ae08417f9ed28edb59a0e7b