keep awaitable around
Summary:
The `Awaitable` type is not kept by hackc, so a function returning
`Awaitable<Foo>` is emitted as returning `Foo`. The full user source annotation is kept around as (interned) string.

This diff keeps around the information that `Awaitable` was present as a Textual type attribute, by doing minimal processing on the user source type.

Note that if someone alias the `Awaitable` type, it might not work. But there's no way the compiler can deal with such type aliases in a systematic way since the alias might be defined in another file.

Reviewed By: davidpichardie

Differential Revision: D51200367

fbshipit-source-id: a9525b8fdf8b3a5433f4f3d72f9db6b337b1232c