Use MADV_WILLNEED to prefetch the .zhhdg during compression
Summary:
Several users with devbigs are reporting pathologically slow .zhhdg
decompression times (5 minutes instead of 10 seconds). We don't know
why it happens, but one theory involves reading from multiple sections
in a file on a spinning disk in parallel, perhaps where a compressed
filesystem is involved.

As an experimental fix, use `madvise` with `MADV_WILLNEED` to tell the
OS that we are going to want all of the input file contents, so it
can start prefetching.

Reviewed By: edwinsmith

Differential Revision: D51208096

fbshipit-source-id: 4cc869e0404bd73c0c6a1bd94ffbc5965350e291