use a 1 MB output buffer when decompressing .zhhdg
Summary:
Several users with devbigs are reporting pathologically slow .zhhdg
decompression times (5 minutes instead of 10 seconds). We don't know
why it happens, but one theory involves writing to multiple sections
of a file on a spinning disk in parallel, perhaps where a compressed
filesystem is involved.

This raises the BufWriter size of the only write in a loop from
8 KiB to 1 MiB, to see if it helps. As the output is > 1 GiB in
size, and we are writing values stored in memory, the extra buffer
memory use is de minimis.

Reviewed By: edwinsmith

Differential Revision: D51208229

fbshipit-source-id: 2f72e318a111df8468c5fbe8d68b92f5011114d7