Remove locally defined partially applied 'helpers' for recursive calls
Summary:
This diff is part of a stack to remove the collection of optional params used in typing to represent:
i) Context
ii) Configuration

The diff eliminates the use of helper function where some of the context / config args are partially applied to a recursive call. This means that all config and context that is passed is made explicit at the call site making subsequent refactors less error prone.

Reviewed By: CatherineGasnier

Differential Revision: D51202224

fbshipit-source-id: d5b4c634622018988989d2ba20ec39771aff768a