Ban InterpOne from EndCatch teardown optimization
Summary:
Some interpreted instruction modify stack before throwing. This is generally
bad, as it conflicts with what the catch block for InterpOne might want to do
and how it might or might not sync the stack.

It doesn't only result in undefined behavior, but also breaks internal
invariants of the EndCatch teardown optimization.

Let's ban it from there until we can fix the behavior of interpreter.

Reviewed By: mdko

Differential Revision: D51268066

fbshipit-source-id: 5429870b088f38cfe2fd5203c1800cb54e5e9f29