replace FuncCounterMap with a byte in each Func
Summary:
We previously count the number of JIT attempts for each function using a tbb hashmap, which can cause heavy contention. Alternative hash table implementations didn't work well. So we find a free byte in Func and do it there.
The downside of it is that the threshold can be at most 256, which seems fine in perf testing.

Reviewed By: jano

Differential Revision: D51049165

fbshipit-source-id: 8967fa81f5a596898a1beafa56b25583e5d029f0