reconstruct type constant informations
Summary:
During compilation, hackc throws away a lot of information about type constant definitions. This diff tries to compute some of them back.

This diff is not the right way to do things: we should keep the original `hint` around long enough to use it during conversion to textual. However such a change should be discussed while aorenste is here.

In the meantime, I do a partial reconstruction to back some type information at the definition of a type constant

Reviewed By: davidpichardie

Differential Revision: D51256254

fbshipit-source-id: 97a63679d6cc12647ef6d94eec1e549ca28e7ccc