Move si_kind into FileInfo
Summary:
My goal for this diffstack is to redo the way ClientIdeIncremental, Provider_utils.quarantine and Provider_utils.invalidate work. (1) Users currently experience bugs due to them not invalidating the necessary decls. (2) They invalidate unnecessary decls, which I suspect hurts perf.

-------

This diff moves si_kind and si_addendum from SearchTypes.ml into FileInfo.ml, nothing more. (The numerous other files in this diff are just mindless consequence of changes in those two files).

```
type si_kind =
  | SI_Class
  | SI_Interface
  | SI_Enum
  | SI_Trait
  | SI_Unknown
  | SI_Mixed
  | SI_Function
  | SI_Typedef
  | SI_GlobalConstant
  | SI_XHP
  | SI_Namespace
  | SI_ClassMethod
  | SI_Literal
  | SI_ClassConstant
  | SI_Property
  | SI_LocalVariable
  | SI_Keyword
  | SI_Constructor

type si_addendum = {
  sia_name: string;
  sia_kind: si_kind;
  sia_is_abstract: bool;
  sia_is_final: bool;
}
```

I'm doing this because these two types end up being part of the file-info "lingua franca", a common part of the file information that we derive from direct-decl-parser and use to power invalidation. In particular,
1. FileInfo.t, si_addendum, and (in future) pfh_hash come out of the rust-batch-indexer which uses a multithreaded direct-decl-parse.
2. FileInfo.t is used by ClientIdeIncremental to update the naming table
3. si_addendum is used by the search-index to update the local delta symbol database for things like autocomplete
3. FileInfo.t and (in future) pfh_hash are used by Provider_utils to invalidate decls.
4. It is ClientIdeDaemon.ml which mediates production and consumption of the information.

At the moment they can't yet be part of the lingua franca. That's because si_addendum is part of search-service, and it would be wrong to build something as fundamental as Provider_utils on top of a high-level user-facing concept like search-service. Therefore, ClientIdeDaemon and ClientIdeIncremental have to do irritating plumbing work to strip out si_addendum and reformulate the rest into a datatype suitable for Provider_utils.

The plumbing will get more irritating in the future as I (1) add pfh_hash to optimize a common case, (2) make a further optimization in case of just one single file change.

So in this diff I'm taking the stance that si_addenum really represents facts about the decls in a file, very similar to FileInfo.t, and therefore belongs at a low level (FileInfo.ml) alongside it. I think this is fair: FileInfo.t and si_addendum are both things derived from direct-decl-parsing.

The result of this diff is that pretty soon I'll be able to remove boilerplate from ClientIdeDaemon and ClientIdeIncremental, and my planned future changes won't add much complexity to the code either.

Reviewed By: madgen

Differential Revision: D51264208

fbshipit-source-id: 29498e4e7d4e2f7ea5a87f600d90f65c41b55ac4