Unwind pending C++ exceptions intead of Hack exceptions
Summary:
After tearDownFrame(), we might have a new pending C++ exception coming from
reentering the VM via event hook. If that happens and we were unwinding a Hack
exception, replace it with the pending C++ exception.

This improves the behavior, as the request will fatal with the C++ exception
instead of a mysterious unhandled Hack exception despite presence of try/catch,
as demonstrated by the updated test output.

Reviewed By: ricklavoie

Differential Revision: D51321158

fbshipit-source-id: a78a4b7c565571182edc5d11bb4ee5af526b34d5