Reset state of frame ctx when creating block with unprocessed preds
Summary:
When creating a block with unprocessed preds, we might get more incoming edges
from anywhere else within the function after the block was emitted, so we need
to reset various parts of the frame state, such as the known state of locals.

While the frame ctx value doesn't change during the function execution, its
SSATmp might be different, either due to being originally created in different
FuncEntry specializations of closures, or due to type refinements. It is
incorrect to reuse whatever refinement we have at the time of creation of
a block with unprocessed preds.

Fix this by keeping track of the original unrefined ctx type (which is the same
across all FuncEntry specializations), then reset ctx and set ctxType to this
type in blocks with unprocessed preds.

Reviewed By: ricklavoie

Differential Revision: D51324947

fbshipit-source-id: ab34c316833a62286cd5676015bc69470e2ec3d7