Show the int part or Local_id.t in pretty printer
Summary:
With TAST hashes, we're better at detecting TAST differences. It's therefore important that `Tast.pp` shows all the details of the TAST. This diff is one step in that direction.

Local_id.t is currently `int * string`. So far, the pretty printer would ignore the int. To avoid having to fix too many tests, we now still skip the int when it's zero (the most common case), but print it when it's not zero.

Reviewed By: madgen

Differential Revision: D51259329

fbshipit-source-id: 101aa305878614edfade9ee2e7adb99a77e5995e