IDE-init is fine with name collisions
Summary:
This diff concerns what happens when you start the IDE in a repository with duplicate names and no saved-state (hence, for OSS, and for non-WWW directories like simple_repo and the directory that test_lsp.py uses).

In the past, duplicate names used to be weird! But we've cleaned them up now:
1. the naming table stores the winner
2. the IDE always respects winner

Therefore, there's no reason to disallow duplicate names. Everything works just fine.

There will be no impact on WWW.

Reviewed By: nt591

Differential Revision: D51351301

fbshipit-source-id: aba08b909d7ee2658d15914043101cf0c618e2c4