make hh_show not break ClientLsp
Summary:
The previous diff made it so when we open test/typecheck files in VSCode then we get IDE services.

I discovered that if I open a file that contains `hh_show` then it crashes the IDE! That's because hh_show prints to stdout, but stdout must be reserved solely for the JSONRPC stream VSCode detects errors in the JSONRPC stream and shuts it down.

Reviewed By: mheiber

Differential Revision: D51351302

fbshipit-source-id: 1e62fd44b33b3ce29a41447145910a32f5de6600