Do not generate `Hany` during `Hvec_or_dict` elaboration
Summary:
We currently generate errors warning that `vec_or_dict` is applied to too many or too few type arguments but will generate a `Hany` in the case of too few.

This diff leaves the misapplied type constructor unelaborated in both cases.

This should only change the error generated (to a `Unbound name`) error should the naming error ever be suppressed.

Reviewed By: andrewjkennedy

Differential Revision: D51394024

fbshipit-source-id: 066a135999071d10040f05c855b6c9090858fc8e