Fix crash in XLS when using phys registers in phijmp
Summary:
analyzePhiHints() builds phi_groups that contains list of all input registers
of a phi for the purpose of hinting register allocation of a phi dst.

This logic assumed that phijmp uses do not contain phys registers. However, in
certain situations, this phijmp might receive a phys register, such as when
using PtrToLval to the top stack element in resumed functions get simplified to
%rbx.

Let's ignore phys register inputs. It is okay to ignore, as it is only a hint
and it would be wrong to try to use the same phys reg for a phi.

Reviewed By: paulbiss

Differential Revision: D51415539

fbshipit-source-id: 5e575d2240bf8bc99876d0dc6913c75cb34ea31e