sticky_quarantine
Summary:
My goal for this diffstack is to redo the way ClientIdeIncremental, Provider_utils.quarantine and Provider_utils.invalidate work. (1) Users currently experience bugs due to them not invalidating the necessary decls. (2) They invalidate unnecessary decls, which I suspect hurts perf.

---------------

This diff changes the way quarantine+invalidation work. We enter quarantine every single time we compute a TAST due to modified source text (i.e. after each DidOpen, DidChange, and also for Autocomplete because it modifies the source text with AUTO332). This diff I hope will make quarantine faster in the common case that decl signatures in the current file haven't changed. My hope is that by recognizing and fast-tracking this common case, we'll have more leeway to do the correct but expensive invalidation for cases where decl signatures have changed.

Prior to this diff,
* **The decl-invariant**: all decls present in the cache (both shallow and folded), and all cached TASTs, reflect the truth as of disk content
* **Quarantine-invariant**: except, for the duration of quarantine, they instead reflect a combination of truth in the *entries* with which we quarantined, combined with the truth as of disk content for everything else.
* **Implementation**. Note that if you deleted every decl from cache and every cached TAST then it would trivially satisfy the decl-invariant and the quarantine-invariant! Our implementation of quarantine didn't quite go that far. It merely deleted affected cache items upon start of quarantine so that, during TAST-computation, if ever we need a decl or folded-decl then it will be re-populated with quarantine-entry-specific decls; this is sufficient to satisfy the quarantine-invariant. And it also deletes them upon exit of quarantine; this is sufficient to restore the decl-invariant.
   * Although, it didn't calculate "affected folded decls" correctly, because doing it correctly made things too slow.

This diff allows for a kind of "sticky quarantine". The idea is that for a single keystroke there's no point invalidating all those decls at the start of autocomplete, and the end of autocomplete, at the start of didchange, and at the end of didchange, when the user was probably typing in a method body and no decls actually had any material changes. Instead, it lets the last-used quarantine just carry over into future compatible quarantines.
```
  decls_reflect_this_file:
    (Relative_path.t * FileInfo.t * FileInfo.pfh_hash) option ref;
```
After this diff, we have different invariants:
* **The decl-invariant**. All decls present in the cache, and all cached TASTs, reflect the truth as of `decls_reflect_this_file` if present, and for all other files reflect the truth of what's on disk.
* **Quarantine-invariant**. Except, for the duration of quarantine, they as before reflect a combination of truth in the *entries* with which we quarantined, combined with the previous truth for everything else.
* Observation: In the common case where we enter quarantine with respect to a single entry, and that entry's `path, pfh_hash` is identical to `decls_reflect_this_file`, then quarantine-entry is a no-op! the decl-invariant and the quarantine-invariant are both trivially satisfied. Also, quarantine-exit is a no-op too!

-------

My first implementation of this broke the test `ide_hierarchy_file_change_on_disk`. The scenario there was
```
/// entry                  /// disk-file
class D extends B {        class B {
}                            function foo(): int {}
                           }
$d->foo();
                           /// modified to:
                             function foo(): string {}
```
Before this diff:
1. We entered quarantine for `entry` and compute TAST. After computation, cache contains folded decls for B and D
2. We exited quarantine. Cache solely contains shallow and folded decl for B.
3. We modified the disk-file. We should have invalidated all affected derived facts, but that was too slow, so we only invalidated the shallow and folded decls of B.
4. We entered quarantine a second time for `entry`. This didn't do anything since folded+shallow decls of D were already absent.
5. We computed TAST for `entry`. This calculated shallow+derived decls for D. In order to do this it had to reconstruct shallow+folded decls for B too, which had been removed by step 3. Our new decls for B reflect the new disk, and hence so do the new decls for D.

After v1 of this diff:
1. We entered quarantine for `entry` and compute TAST. After computation, cache contains folded decls for B and D, and `decls_reflect_this_file` reflect `entry`.
2. We exited quarantine. Cache was left intact.
3. We modified the disk-file. Again we invalidated shallow and folded decls of B.
4. We entered quarantine a second time for `entry`. The `pfh_hash` of `entry` is the same as it was last time, so we don't invalidate any decls.
5. We computed TAST for `entry`. This is now using the wrong decl for `D`.

In effect, prior to this diff, the excessive invalidation that we did upon quarantine entry+exit was papering over (in this scenario) the insufficient invalidation we did for disk-changes.

Therefore v2 of this diff changed the disk-modified invalidation too. It is now correct: upon disk change we will invalidate all folded decls. (but, see the next diff in the stack for an optimization in the common case).

In effect this diff is a gamble. I hope that by being cheap in the common case, it gives us freedom to be expensive+correct in the rare cases. We'll see when it rolls out whether the gamble worked.

Reviewed By: mheiber

Differential Revision: D51224222

fbshipit-source-id: 1658300c5d2b771fc556905f2dfbc0bace9fca7f