avoid invalidating if user saves the file they have open
Summary:
Earlier in this diff-stack we made a gamble:
1. Make the common case of "user makes a change that doesn't affect decls" cheap
2. Make the case of "user changes a disk file" correct at the cost of being expensive
3. (I still haven't changed it so "user makes a change that does effect decls" is expensive+correct; I've retained the existing incorrect+fast invalidation. I plan to do this in a future experiment, under a new JK)

I figured there's an easy+clean optimization to make in what I suspect is the common case, where a user presses Cmd+S to save the file they currently have open. Some users save quite frequently. In this case there's no need to invalidate any decls, since the disk content will be just the same as the editor content.

That's what this diff does! If the new disk content is equal to `decls_reflect_this_file`, then the decl-invariant is already satisfied, and we don't need to invalidate any decls.

I might be doing premature optimization? But it's an easy one, and I'd rather start rolling out my best stab at the JK.

Reviewed By: mheiber

Differential Revision: D51292216

fbshipit-source-id: 1c1b7c66c2ad9b23297a2cc1c82166dcef4a0dc4