Cache results of top inputs in aggregate_data
Summary:
Store `aggregateData` in `LocalIndex`.

- Cache the result of `aggregate_data` for top inputs and re-use on subsequent invocations
- This diff is strictly better, since currently we always calculate all top nodes, so if we find a top node while calculating
another top node, that top node was either already processed and we find the result in our cache, or we compute it now to reuse later. In addition, sorting transitive children by top nodes first, means that we will not have any redundant computations for top nodes.

This is certainly not optimal.

1) We don't cache any non-top node, even when multiple top nodes need it. To handle these cases, we would need some sort of
"reaching tops" analysis to ensure we are not duplicating any computations. I'm brainstorming some better algorithms,
but for now, this should be better.

Some observations:
1. We will get no benefit from this in round 1 where we process 1 top input per job.
2. Batching related classes would benefit much more from this change.

Reviewed By: ricklavoie

Differential Revision: D51285601

fbshipit-source-id: 502e379dcd927b05a671a9dba9e6f9cf4a8c2a6a