fix dedupe logic for nonstandard types (unbreak hargow server)
Summary:
the recently landed D51368304  broke the hargow server build.

D51368304 searches through typedefs for rust nonstandard type annotations. it collects the nonstandard types found in these annotations into a set so that serialization code for them can be emitted.

the comparator for the set is used to avoid inserting the same type twice. the comparator logic wasn't updated in D51368304. this led to types that should be there being left out of as they were getting erroneously classified as already present.

Reviewed By: wtetzner

Differential Revision: D51445302

fbshipit-source-id: e6c961fea1a536fbeaebae84f0f319c7de31c346