Remove top level exception handlers with UnsetLs
Summary:
The frontend often generates catch blocks that UnsetL temporary unnamed locals.
This is useful when there is a try/catch around such code to keep liveness of
temporary state in check. But it doesn't serve any purpose when the exception
is not going to be handled and the whole frame is going to be unwound.

Remove such exception handlers in hhbbc.

Reviewed By: ricklavoie

Differential Revision: D14682187

fbshipit-source-id: ab504f0c4a7caeab5e923fbce1908f7ca373ddbb