Decl_enforcability should call get_typedef_without_pessimise
Summary:
D39314187 (Sep 2022) made it so decl_enforceability invoked a "get_typedef" function that was specific solely to shmem (not rust-provider as used in hh_server, not localmem as used in IDE).

Therefore, every time decl_enforceability needed a typedef, it first failed to find it in `Decl_store.get()` (since rust-provider and localmem never write into Decl_store). Then it took the slightly slower route of looking up naming-provider, and obtaining the direct-decl-parsed results. Thankfully they were usually cached.

This diff corrects it. The real goal was that decl_enforceability should call a version of get_typedef which doesn't do pessimisation (since that would be a circular dependency). It's clear now after this diff what it does.

Reviewed By: andrewjkennedy

Differential Revision: D51442098

fbshipit-source-id: 73cfa53a689b72b85fb1470beac4e478157fa20e