delete hh_rearch
Summary:
The hack re-architecture didn't succeed. This diff therefore deletes it.

(One thing that's exciting -- with this diff, coupled with the previous one that deletes hh_typed_pessimisation, we get the number of Provider_backends down from 6 to 4. I wonder if we can also delete the Shared_memory backend? that would leave only 3 providers -- Rust for hh_server and tools, Localmem for IDE, and AI for zoncolan.)

This diff also deletes the sandcastle_perf.sh flag which runs the re-architecture. A separate diff D51450837 deletes the fbpkg definitions associated with the re-architecture.

Reviewed By: edwinsmith

Differential Revision: D51450649

fbshipit-source-id: 0a5a2d9d6f35b28c0cee985f38c1ef4f5d3330a0