configurable eden_fetch_parallelism
Summary:
This adds support for `eden_fetch_parallelism` solely for the `check` step
and Justknobs options as well, so we can pass it along via JK
or override with `--config` keys.

This brings observed filesystem spidering down from 180 seconds to 30 seconds.
I don't know the P99, so we'll have to roll out and measure safely.

Reviewed By: edwinsmith

Differential Revision: D51428057

fbshipit-source-id: b517ac25e3b1e1bbb381532642bda0d94d0de094