Optimized parent fanout: handle files with errors
Summary:
The algorithm introduced in the previous diff is meant to discover new errors, but not determine which existing errors can be fixed.

One way around that is to recheck all files with errors. This can be problematic if there are many errors, for example, with a codemod. Rechecking all files with errors each time the user makes a fix will guarantee long iterations. So to stay on the safe side, the get_fanout function returns two deps sets: one to always recheck, and one to recheck only if the associated files previously had errors. These two sets are encapsulated in type `Fanout.t`.

Differential Revision: D51025392

fbshipit-source-id: 97be6267255940d1162c79568459f98374653909