Fanout: consolidate handling of files with errors.
Summary:
In the previous diff, we changed how to include files with errors in the fanout.
It turns out there is another place in the code which includes some files with errors in the fanout, based on whether the positions in the errors are obsolete. In this diff, we simply move that logic closer to the logic introduced in the previous diff, so a reader of the code gets a complete picture at once.

Differential Revision: D51027243

fbshipit-source-id: d456ef4b046fe758470299b970e0db8590a72846